Non metric multidimensional scaling

A wrapper around vegan::metaMDS.

NMDS(x, distance = "bray", k = 2, try = 20, trymax = 20, ...)

Arguments

x
any Coe object

distance
a dissiminarity index to feed vegan::vegdist (default: bray)

k
numeric number of dimensions to feed vegan::metaMDS (default: 2)

try
numeric minimum number of random starts to feed vegan::metaMDS (default: 20)

trymax
numeric minimum number of random starts to feed vegan::metaMDS (default: 20)

...
additional parameters to feed vegan::metaMDS

Value

what is returned by vegan::metaMDS plus $fac. And prepend NMDS class to it.

Details

For Details, see vegan::metaMDS

See also

Examples

x <- bot %>% efourier %>% NMDS
#> 'norm=TRUE' is used and this may be troublesome. See ?efourier
#> 'nb.h' set to 10 (99% harmonic power)
#> Registered S3 method overwritten by 'vegan':
#>   method     from      
#>   rev.hclust dendextend
#> 'comm' has negative data: 'autotransform', 'noshare' and 'wascores' set to FALSE
#> Warning: results may be meaningless because data have negative entries in method “bray”
#> Run 0 stress 0.07227125 
#> Run 1 stress 0.07227125 
#> ... New best solution
#> ... Procrustes: rmse 3.936045e-06  max resid 1.87576e-05 
#> ... Similar to previous best
#> Run 2 stress 0.1536609 
#> Run 3 stress 0.1729702 
#> Run 4 stress 0.07227125 
#> ... Procrustes: rmse 5.981988e-06  max resid 2.874886e-05 
#> ... Similar to previous best
#> Run 5 stress 0.07227125 
#> ... Procrustes: rmse 2.19081e-06  max resid 8.275813e-06 
#> ... Similar to previous best
#> Run 6 stress 0.07227125 
#> ... Procrustes: rmse 7.028881e-06  max resid 2.808569e-05 
#> ... Similar to previous best
#> Run 7 stress 0.07227125 
#> ... Procrustes: rmse 9.669053e-06  max resid 3.852924e-05 
#> ... Similar to previous best
#> Run 8 stress 0.1476475 
#> Run 9 stress 0.07227125 
#> ... Procrustes: rmse 1.146136e-06  max resid 4.040052e-06 
#> ... Similar to previous best
#> Run 10 stress 0.07227125 
#> ... Procrustes: rmse 2.035531e-06  max resid 8.084339e-06 
#> ... Similar to previous best
#> Run 11 stress 0.07227126 
#> ... Procrustes: rmse 7.291301e-06  max resid 2.685738e-05 
#> ... Similar to previous best
#> Run 12 stress 0.1647891 
#> Run 13 stress 0.07227125 
#> ... Procrustes: rmse 6.99825e-06  max resid 2.725125e-05 
#> ... Similar to previous best
#> Run 14 stress 0.07227125 
#> ... Procrustes: rmse 5.714743e-06  max resid 2.216745e-05 
#> ... Similar to previous best
#> Run 15 stress 0.1606935 
#> Run 16 stress 0.07227125 
#> ... Procrustes: rmse 6.971937e-06  max resid 2.720148e-05 
#> ... Similar to previous best
#> Run 17 stress 0.07227125 
#> ... Procrustes: rmse 8.281514e-07  max resid 2.450366e-06 
#> ... Similar to previous best
#> Run 18 stress 0.1776127 
#> Run 19 stress 0.07227125 
#> ... Procrustes: rmse 4.920735e-06  max resid 1.945613e-05 
#> ... Similar to previous best
#> Run 20 stress 0.07227125 
#> ... Procrustes: rmse 1.95309e-06  max resid 6.877957e-06 
#> ... Similar to previous best
#> *** Solution reached

# Shepard diagram # before a Momocs wrapper
# vegan::stressplot(x)