Retains group with at least a certain number of individuals

Examples are self-speaking.

at_least(x, fac, N)

Arguments

x
any Momocs object

fac
the id of name of the $fac column

N
minimal number of individuals to retain the group

Note

if N is too ambitious the original object is returned with a message

See also

Other handling functions: arrange, chop, combine, dissolve, filter, mutate, rename, rescale, rm_harm, rm_uncomplete, rw_fac, sample_frac, sample_n, select, slice, subset.Coo, table

Examples

data(trilo)
table(trilo, "onto")

#> 
#>  a  b  c  d 
#>  7 16 18  9 

at_least(trilo, "onto", 9)

#> An Out object with: 
#>  - $coo: 43 outlines (64 +/- 0 coordinates, all unclosed)
#>  - $fac: 1 classifier:
#>      'onto' (factor 3): b, c, d.

at_least(trilo, "onto", 16)

#> An Out object with: 
#>  - $coo: 34 outlines (64 +/- 0 coordinates, all unclosed)
#>  - $fac: 1 classifier:
#>      'onto' (factor 2): b, c.

at_least(trilo, "onto", 2000) # too ambitious !

#> no group with at least 2000 indidivuals

#> An empty Out object