Calculate classification metrics on a confusion matrix

In some cases, the class correctness or the proportion of correctly classified individuals is not enough, so here are more detailed metrics when working on classification.

classification_metrics(x)

Arguments

x
a table or an LDA object

Value

a list with the following components is returned:

accuracy the fraction of instances that are correctly classified

macro_prf data.frame containing precision (the fraction of correct predictions for a certain class); recall, the fraction of instances of a class that were correctly predicted; f1 the harmonic mean (or a weighted average) of precision and recall.

macro_avg, just the average of the three macro_prf indices

ova a list of one-vs-all confusion matrices for each class

ova_sum a single of all ova matrices

kappa measure of agreement between the predictions and the actual labels

See also

The pages below are of great interest to understand these metrics. The code used is partley derived from the Revolution Analytics blog post (with their authorization). Thanks to them!

https://en.wikipedia.org/wiki/Precision_and_recall

http://blog.revolutionanalytics.com/2016/03/com_class_eval_metrics_r.html

http://www.r-bloggers.com/is-your-classification-model-making-lucky-guesses/

Other multivariate: CLUST, KMEANS, LDA, MANOVA_PW, MANOVA, PCA, mshapes

Examples

# some morphometrics on 'hearts'
hearts %>% fgProcrustes(tol=1) %>%
coo_slide(ldk=1) %>% efourier(norm=FALSE) %>% PCA() %>%
# now the LDA and its summary
LDA(~aut) %>% classification_metrics()

#> 'nb.h' not provided and set to 7 (99% harmonic power)

#> 0.99 total variance

#> 10 PC retained

#> $accuracy
#> [1] 0.775
#> 
#> $macro_prf
#>       precision    recall        f1
#> ced   0.8666667 0.8666667 0.8666667
#> jeya  0.7500000 0.8000000 0.7741935
#> mat   0.5833333 0.4666667 0.5185185
#> ponnu 0.9259259 0.8333333 0.8771930
#> remi  0.8620690 0.8333333 0.8474576
#> rom   0.8387097 0.8666667 0.8524590
#> ruks  0.8709677 0.9000000 0.8852459
#> vince 0.5277778 0.6333333 0.5757576
#> 
#> $macro_avg
#>   avg_precision avg_recall    avg_f1
#> 1     0.7781813      0.775 0.7746865
#> 
#> $ova
#> $ova$ced
#>         classified
#> actual   ced others
#>   ced     26      4
#>   others   4    206
#> 
#> $ova$jeya
#>         classified
#> actual   jeya others
#>   jeya     24      6
#>   others    8    202
#> 
#> $ova$mat
#>         classified
#> actual   mat others
#>   mat     14     16
#>   others  10    200
#> 
#> $ova$ponnu
#>         classified
#> actual   ponnu others
#>   ponnu     25      5
#>   others     2    208
#> 
#> $ova$remi
#>         classified
#> actual   remi others
#>   remi     25      5
#>   others    4    206
#> 
#> $ova$rom
#>         classified
#> actual   rom others
#>   rom     26      4
#>   others   5    205
#> 
#> $ova$ruks
#>         classified
#> actual   ruks others
#>   ruks     27      3
#>   others    4    206
#> 
#> $ova$vince
#>         classified
#> actual   vince others
#>   vince     19     11
#>   others    17    193
#> 
#> 
#> $ova_sum
#>           classified
#> actual     relevant others
#>   relevant      186     54
#>   others         54   1626
#> 
#> $kappa
#> [1] 0.7428571
#>