Classify using LDA

Classify using LDA

classify(x, fac, ref, unk)

Arguments

x
a Coe

fac
a standalone factor, or the name or id of the $fac column to use.If it contains NAs, they will also be removed first from the x object

ref
at least two level names from $fac to use as a training subset of x

unk
same as above for one level name to classify

Value

a list with components:

$N_ref the number of elements in the training set

$N_unk the number of elements in the unknown set

$counts counts of classification of 'unk' in each class of 'ref'

$pc same thing as percentages

$probs same thing as posterior probabilities

$probs same thing as posterior but as a data.frame

Examples

table(olea, "var")

#> 
#>  Aglan  Cypre MouBo1  PicMa 
#>     60     30     60     60 

x <- opoly(olea, 5, verbose=FALSE)

#> 'nb.pts' missing and set to 91

classify(x, fac="var", ref=c("Aglan","Cypre"), unk="PicMa")

#> $N_ref
#> [1] 90
#> 
#> $N_ref_tab
#> 
#> Aglan Cypre 
#>    60    30 
#> 
#> $N_unk
#> [1] 60
#> 
#> $counts
#> 
#> Aglan Cypre 
#>    56     4 
#> 
#> $pc
#> 
#> Aglan Cypre 
#> 93.33  6.67 
#> 
#> $probs
#>                       Aglan       Cypre
#> 0023-cPicMa_O10VD 0.8172552 0.182744835
#> 0023-cPicMa_O10VL 0.9185876 0.081412404
#> 0023-cPicMa_O11VD 0.9116187 0.088381297
#> 0023-cPicMa_O11VL 0.9193335 0.080666508
#> 0023-cPicMa_O12VD 0.9820535 0.017946451
#> 0023-cPicMa_O12VL 0.9533485 0.046651461
#> 0023-cPicMa_O13VD 0.5623658 0.437634229
#> 0023-cPicMa_O13VL 0.9036465 0.096353550
#> 0023-cPicMa_O14VD 0.8714521 0.128547879
#> 0023-cPicMa_O14VL 0.9792518 0.020748207
#> 0023-cPicMa_O15VD 0.8946710 0.105329044
#> 0023-cPicMa_O15VL 0.9273385 0.072661483
#> 0023-cPicMa_O16VD 0.8144305 0.185569482
#> 0023-cPicMa_O16VL 0.8878878 0.112112173
#> 0023-cPicMa_O17VD 0.8743407 0.125659261
#> 0023-cPicMa_O17VL 0.9556530 0.044346988
#> 0023-cPicMa_O18VD 0.6551452 0.344854762
#> 0023-cPicMa_O18VL 0.9051242 0.094875825
#> 0023-cPicMa_O19VD 0.9354976 0.064502416
#> 0023-cPicMa_O19VL 0.4556499 0.544350095
#> 0023-cPicMa_O1VD  0.9799853 0.020014650
#> 0023-cPicMa_O1VL  0.9850076 0.014992426
#> 0023-cPicMa_O20VD 0.8247911 0.175208928
#> 0023-cPicMa_O20VL 0.1593337 0.840666302
#> 0023-cPicMa_O21VD 0.9523456 0.047654434
#> 0023-cPicMa_O21VL 0.8012307 0.198769299
#> 0023-cPicMa_O22VD 0.8273051 0.172694919
#> 0023-cPicMa_O22VL 0.9429347 0.057065261
#> 0023-cPicMa_O23VD 0.6598153 0.340184689
#> 0023-cPicMa_O23VL 0.9610699 0.038930072
#> 0023-cPicMa_O24VD 0.6370924 0.362907553
#> 0023-cPicMa_O24VL 0.9825320 0.017467954
#> 0023-cPicMa_O25VD 0.7849479 0.215052113
#> 0023-cPicMa_O25VL 0.9893202 0.010679755
#> 0023-cPicMa_O26VD 0.3572978 0.642702161
#> 0023-cPicMa_O26VL 0.9836236 0.016376359
#> 0023-cPicMa_O27VD 0.8840060 0.115994005
#> 0023-cPicMa_O27VL 0.9397928 0.060207244
#> 0023-cPicMa_O28VD 0.3525263 0.647473727
#> 0023-cPicMa_O28VL 0.8772828 0.122717193
#> 0023-cPicMa_O29VD 0.7390772 0.260922791
#> 0023-cPicMa_O29VL 0.8480262 0.151973846
#> 0023-cPicMa_O2VD  0.9311705 0.068829495
#> 0023-cPicMa_O2VL  0.9591539 0.040846126
#> 0023-cPicMa_O30VD 0.8692837 0.130716349
#> 0023-cPicMa_O30VL 0.9911134 0.008886561
#> 0023-cPicMa_O3VD  0.5880704 0.411929640
#> 0023-cPicMa_O3VL  0.9687164 0.031283642
#> 0023-cPicMa_O4VD  0.9587286 0.041271360
#> 0023-cPicMa_O4VL  0.9668060 0.033193957
#> 0023-cPicMa_O5VD  0.9174841 0.082515938
#> 0023-cPicMa_O5VL  0.9864513 0.013548697
#> 0023-cPicMa_O6VD  0.8014485 0.198551543
#> 0023-cPicMa_O6VL  0.9560072 0.043992810
#> 0023-cPicMa_O7VD  0.9711144 0.028885599
#> 0023-cPicMa_O7VL  0.9892752 0.010724840
#> 0023-cPicMa_O8VD  0.8310649 0.168935111
#> 0023-cPicMa_O8VL  0.6619930 0.338006968
#> 0023-cPicMa_O9VD  0.6302951 0.369704935
#> 0023-cPicMa_O9VL  0.9714413 0.028558711
#> 
#> $probs_fac
#> NULL
#>