Helps to select a given number of harmonics from a numerical vector.

coeff_sel helps to select a given number of harmonics by returning their indices when arranged as a numeric vector. For instance, harmonic coefficients are arranged in the $coe slot of Coe-objects in that way: \(A_1, \dots, A_n, B_1, \dots, B_n, C_1, \dots, C_n, D_1, \dots, D-n\) after an elliptical Fourier analysis (see efourier and efourier) while \(C_n and D_n\) harmonic are absent for radii variation and tangent angle approaches (see rfourier and tfourier respectively). . This function is used internally but might be of interest elwewhere.

coeff_sel(retain = 8, drop = 0, nb.h = 32, cph = 4)

Arguments

retain
numeric. The number of harmonics to retain.

drop
numeric. The number of harmonics to drop

nb.h
numeric. The maximum harmonic rank.

cph
numeric. Must be set to 2 for rfourier and tfourier were used.

Value

coeff_sel returns indices that can be used to select columns from an harmonic coefficient matrix. coeff_split returns a named list of coordinates.

Examples

bot.f <- efourier(bot, 32)

#> you selected `norm=TRUE`, which is not recommended. See ?efourier

coe <- bot.f$coe # the raw matrix
coe

#>                A1           A2         A3         A4          A5          A6
#> brahma          1  0.006766531 0.09348184 0.01374288 0.023818571 0.008592275
#> caney           1  0.007368844 0.09542847 0.01449765 0.022207326 0.006665955
#> chimay          1  0.016535135 0.07571705 0.02736989 0.009876978 0.014306122
#> corona          1  0.007486279 0.09616977 0.01220177 0.022887407 0.007281456
#> deusventrue     1  0.019627538 0.09281761 0.02230953 0.015264921 0.014857953
#> duvel           1  0.022177771 0.06895598 0.03354560 0.009316683 0.021131541
#> franziskaner    1  0.007314669 0.09110003 0.01294912 0.023136784 0.010558958
#>                           A7          A8            A9           A10
#> brahma          0.0031835706 0.005158502 -7.262824e-04  0.0047287291
#> caney           0.0035527091 0.007010166  1.214949e-03  0.0038734169
#> chimay         -0.0047412879 0.007814037 -2.112661e-03  0.0022043011
#> corona          0.0055045888 0.007852411  8.767189e-04  0.0044201528
#> deusventrue     0.0025214510 0.011391904 -1.733965e-03  0.0062083192
#> duvel          -0.0016871288 0.011025502 -1.042906e-04  0.0017042044
#> franziskaner    0.0045813399 0.006927102 -5.793922e-04  0.0045217266
#>                          A11           A12           A13           A14
#> brahma         -0.0013733386  0.0016363823 -2.308262e-03 -8.638239e-05
#> caney          -0.0018777601  0.0011330555 -6.140388e-04  9.095865e-04
#> chimay         -0.0012371979 -0.0018514226  4.671292e-04 -2.460672e-03
#> corona         -0.0021742849  0.0026300698 -1.040704e-03  1.308265e-03
#> deusventrue    -0.0007936789  0.0028680575 -3.667224e-05  4.628731e-04
#> duvel          -0.0006045816 -0.0028942242  1.911167e-03 -1.075121e-03
#> franziskaner   -0.0015947508  0.0021935018 -1.336804e-03  4.860914e-04
#>                          A15           A16           A17           A18
#> brahma         -1.150816e-03 -3.765755e-04 -8.576746e-04 -0.0011152905
#> caney          -4.470877e-04 -9.307845e-04 -6.102760e-04 -0.0008460017
#> chimay          1.560010e-05 -2.625393e-03 -6.111634e-04 -0.0014908027
#> corona         -1.317478e-03 -5.180094e-04 -7.007916e-04 -0.0006290111
#> deusventrue     9.538496e-04 -6.759098e-04  7.618085e-04 -0.0011834834
#> duvel           2.113421e-03 -1.342867e-03 -5.580321e-04 -0.0007953381
#> franziskaner   -8.169832e-04 -7.064707e-04 -4.330142e-04 -0.0013097566
#>                          A19           A20           A21           A22
#> brahma         -8.360520e-04 -1.095155e-03 -5.439578e-04 -0.0008608038
#> caney           2.301119e-05 -8.399357e-04 -5.178245e-04 -0.0009469597
#> chimay         -1.119706e-03 -9.775433e-04 -1.256788e-03 -0.0002828775
#> corona         -3.339722e-04 -1.281244e-03 -4.441549e-04 -0.0010768982
#> deusventrue     4.696431e-04 -8.906751e-04 -3.783424e-05 -0.0004364503
#> duvel          -1.052150e-03  1.767706e-04 -1.315824e-03  0.0001575196
#> franziskaner   -2.487242e-04 -1.440287e-03 -2.906736e-04 -0.0013413124
#>                          A23           A24           A25           A26
#> brahma         -5.735915e-04 -7.299627e-04 -5.033864e-04 -4.409200e-04
#> caney          -3.939390e-04 -4.249811e-04 -3.278067e-04 -2.777037e-04
#> chimay         -6.410642e-04 -3.497060e-04 -3.987327e-04 -3.797373e-04
#> corona         -1.717514e-04 -8.503422e-04 -5.002657e-04 -6.613581e-04
#> deusventrue    -2.719114e-04 -1.349406e-04 -4.678471e-04 -6.742999e-05
#> duvel          -7.818289e-04  3.932732e-04 -9.314404e-06 -2.825668e-05
#> franziskaner   -4.046305e-04 -1.021761e-03 -4.715859e-04 -6.988167e-04
#>                          A27           A28           A29           A30
#> brahma         -3.652079e-04 -2.976623e-04 -3.220750e-04 -2.405557e-04
#> caney          -4.543725e-04 -1.340038e-04 -1.715288e-04  8.287078e-05
#> chimay          1.897326e-04 -3.220192e-04  1.012090e-04 -3.939649e-04
#> corona         -5.628453e-04 -2.866408e-04 -6.301550e-04 -2.277373e-04
#> deusventrue    -4.081023e-04 -6.793693e-06 -2.278672e-04 -1.602607e-04
#> duvel          -4.501744e-05 -5.229843e-04  8.047695e-05 -3.443646e-04
#> franziskaner   -4.854885e-04 -4.750621e-04 -4.823508e-04 -3.278401e-04
#>                          A31           A32 B1            B2            B3
#> brahma         -1.974790e-04 -8.094538e-05  0 -1.900652e-04  3.306231e-04
#> caney          -5.975848e-05 -7.681466e-06  0  5.012013e-04 -3.851293e-04
#> chimay          3.467138e-05 -3.425039e-05  0  1.843629e-04  4.196107e-04
#> corona         -6.391514e-04 -7.778089e-05  0 -3.586724e-04  1.711055e-05
#> deusventrue    -5.672855e-05 -9.370594e-05  0  1.774985e-04 -8.326845e-05
#> duvel           2.041590e-04 -1.384422e-04  0 -4.198782e-04  7.447638e-05
#> franziskaner   -3.511246e-04 -1.887467e-04  0 -8.367911e-04 -3.508429e-04
#>                           B4            B5            B6            B7
#> brahma         -5.191749e-04  1.067446e-04 -9.218905e-05  5.604686e-07
#> caney           3.333918e-04 -2.899903e-04  7.350207e-05 -4.952054e-04
#> chimay          3.227901e-04 -2.906714e-05  5.573360e-04  1.059517e-04
#> corona         -5.501057e-04 -1.907425e-04 -4.256287e-04 -2.147013e-04
#> deusventrue    -1.403373e-03 -3.240180e-04 -9.330047e-04  6.515692e-04
#> duvel          -6.627095e-04  6.107940e-05 -4.746985e-04  2.450959e-04
#> franziskaner   -6.983186e-04 -1.894199e-04 -5.170165e-04  5.988419e-05
#>                           B8            B9           B10           B11
#> brahma          6.268503e-06 -1.960132e-04  1.334431e-04  4.058288e-05
#> caney           8.536695e-05 -4.143363e-04  1.667420e-04 -3.134942e-04
#> chimay          6.209192e-04  1.441644e-04  1.715223e-04  5.930974e-05
#> corona         -1.931107e-04 -3.048499e-04 -2.065499e-04 -2.244684e-04
#> deusventrue    -8.354423e-04  2.287499e-05 -6.080012e-04  2.971827e-04
#> duvel          -1.676532e-04  1.512500e-04  6.986965e-05  2.318703e-04
#> franziskaner   -3.810484e-04  5.133436e-05 -3.407356e-04  5.681736e-05
#>                          B12           B13           B14           B15
#> brahma         -1.917814e-04 -4.415768e-05 -1.760152e-06  2.542883e-05
#> caney           2.011956e-05 -2.220739e-04  9.710885e-05 -2.929034e-05
#> chimay         -3.197006e-06  3.218682e-04 -2.136764e-04  2.996248e-04
#> corona         -6.091420e-06 -1.616363e-04 -1.409505e-05 -3.036983e-04
#> deusventrue    -2.794108e-04 -2.410553e-04 -5.766037e-06 -1.013081e-04
#> duvel           1.659893e-04 -2.147276e-05 -7.372516e-05 -1.697182e-05
#> franziskaner   -2.140707e-04  1.309236e-04 -1.825897e-05  9.453212e-05
#>                          B16           B17           B18           B19
#> brahma         -1.024642e-04 -8.031982e-05 -7.433041e-05  2.949173e-05
#> caney           1.467454e-04 -1.500259e-05  8.962807e-05  1.017678e-04
#> chimay         -4.160132e-04  2.173693e-04 -2.232850e-04  1.410948e-04
#> corona          1.562544e-04 -1.863667e-04  1.898815e-04 -3.051700e-04
#> deusventrue     2.687631e-04 -5.866990e-05  2.532389e-04  4.683303e-05
#> duvel           8.924870e-05  1.724312e-04  1.497399e-04  1.188058e-04
#> franziskaner    2.076190e-05  5.303781e-05  1.447445e-04  2.813898e-05
#>                          B20           B21           B22           B23
#> brahma         -1.727433e-05  3.737724e-05 -9.462283e-06 -4.267458e-05
#> caney           1.631617e-04  1.508881e-04  1.414211e-04  1.088653e-04
#> chimay         -3.610984e-05 -1.712342e-04  2.100688e-05 -3.082575e-04
#> corona          5.852024e-05 -1.892126e-04  1.439899e-04 -1.701985e-04
#> deusventrue     1.245832e-04  7.313250e-06 -1.170924e-04  1.347852e-04
#> duvel           1.101660e-04 -1.728305e-05  9.542637e-05 -1.277279e-04
#> franziskaner    1.795671e-04  6.095791e-06  1.990743e-04  7.411105e-06
#>                          B24           B25           B26           B27
#> brahma          3.357899e-05  1.692467e-05  1.346898e-04  1.323552e-05
#> caney           8.041007e-05  1.705793e-04  6.811154e-05  6.205877e-05
#> chimay          6.975068e-05 -3.230519e-04  2.772097e-05 -1.930852e-04
#> corona          2.066873e-05 -1.628616e-04  7.895544e-05  9.425535e-06
#> deusventrue    -1.235701e-04  1.790799e-04 -1.724233e-04  1.249888e-04
#> duvel           7.759496e-05 -1.131994e-04  1.749479e-04  5.630411e-05
#> franziskaner    2.096840e-04  5.258363e-05  2.046122e-04  9.602614e-05
#>                          B28           B29           B30           B31
#> brahma          9.806530e-05 -1.691212e-05  1.229728e-04  6.153001e-06
#> caney           4.057485e-05  3.928378e-05  3.848225e-05  5.608283e-05
#> chimay         -2.670005e-05  2.629625e-05 -4.657436e-05  1.626961e-04
#> corona         -1.115351e-05  1.797628e-05 -6.512991e-05  1.430262e-04
#> deusventrue    -1.098118e-04 -7.708499e-06 -1.093334e-04 -1.335296e-04
#> duvel           8.996050e-05 -1.871001e-05 -1.512385e-04 -1.753386e-05
#> franziskaner    1.770497e-04  7.088446e-05  8.138285e-05  5.397908e-05
#>                          B32 C1            C2            C3            C4
#> brahma          1.217078e-04  0 -1.637571e-03 -3.936895e-03  5.408096e-03
#> caney           1.227213e-05  0  1.239828e-03 -2.845651e-04  3.757825e-04
#> chimay         -4.552776e-05  0 -3.757608e-03 -1.797357e-03 -2.127924e-03
#> corona         -4.547231e-05  0 -1.652864e-03  1.573302e-03  4.897281e-04
#> deusventrue     5.762098e-05  0  1.552775e-03  7.706329e-04 -1.416448e-03
#> duvel          -1.615768e-04  0  2.872128e-04 -5.422392e-06 -7.785717e-04
#> franziskaner    5.143834e-05  0 -1.253868e-03  3.476506e-04  1.004621e-03
#>                           C5            C6            C7            C8
#> brahma         -1.259407e-03 -3.994402e-03  3.268582e-03  4.792269e-04
#> caney          -4.017802e-05  4.699805e-04  2.518166e-04  6.300072e-04
#> chimay         -4.663387e-04 -7.424827e-05 -8.096453e-05  9.946667e-04
#> corona          1.867708e-04  6.888736e-04  3.145355e-04  5.189042e-04
#> deusventrue     1.463377e-03  9.055123e-04 -2.834923e-04 -1.443910e-03
#> duvel           3.178998e-04  2.219253e-04  5.438377e-04 -3.518900e-06
#> franziskaner    1.054384e-03  2.433412e-04  9.526672e-04  8.662476e-04
#>                           C9           C10           C11           C12
#> brahma         -1.321861e-03  7.225736e-04  2.052451e-04  1.666890e-04
#> caney           4.837212e-04 -3.443095e-04  1.349160e-04  4.336526e-06
#> chimay         -5.660053e-04  2.292011e-04  3.851391e-04 -2.066153e-05
#> corona          3.419355e-04 -4.427112e-05  5.053095e-04  2.703064e-04
#> deusventrue     5.149210e-05  4.983572e-04 -8.471361e-05  3.773608e-04
#> duvel           4.727181e-05  8.518990e-04  5.267156e-05 -7.761916e-06
#> franziskaner    4.052798e-04  2.307566e-04  5.425481e-04  8.822035e-04
#>                          C13           C14           C15           C16
#> brahma         -7.782826e-05 -1.930624e-05  3.379892e-04 -9.611430e-05
#> caney           5.571557e-05  1.585334e-04 -3.676568e-04 -2.424919e-04
#> chimay          2.069414e-04 -1.059161e-04 -5.895162e-05 -4.070464e-04
#> corona         -3.358855e-05  3.794088e-04  7.431417e-05  7.851392e-04
#> deusventrue    -2.738947e-04  7.547067e-04  2.411713e-04  2.337981e-04
#> duvel          -2.899095e-04  1.153102e-04  4.201454e-04 -9.309399e-05
#> franziskaner    2.887549e-04  1.838704e-04 -1.076209e-04 -8.678411e-05
#>                          C17           C18           C19           C20
#> brahma         -2.462206e-05  4.542898e-05  2.594677e-04  1.484945e-04
#> caney          -3.200717e-04 -2.643818e-04  1.448941e-04 -1.295417e-04
#> chimay          2.615626e-04  1.230016e-04 -2.602910e-04  4.566772e-04
#> corona          1.619221e-04 -1.010082e-04 -9.475160e-05  4.604978e-05
#> deusventrue     4.889298e-04  3.580373e-04  2.948538e-04 -2.065040e-05
#> duvel           2.191614e-04  8.201863e-05  5.854558e-06 -8.940682e-05
#> franziskaner   -8.092666e-05 -6.457841e-05 -2.824789e-04 -3.859311e-04
#>                          C21           C22           C23           C24
#> brahma         -2.357683e-04 -3.255489e-04 -3.888181e-05  1.601136e-04
#> caney           1.953747e-04 -2.101846e-04 -1.492333e-04  3.676434e-04
#> chimay         -2.692597e-04  2.257056e-04 -1.397033e-05 -5.646081e-05
#> corona          1.999724e-04  6.189685e-05 -3.340597e-04 -1.953675e-04
#> deusventrue     1.591834e-04 -2.832760e-04  1.458864e-04 -8.427409e-05
#> duvel          -2.955388e-04 -1.892748e-05 -2.437293e-04 -1.378691e-04
#> franziskaner   -1.525330e-04  1.618032e-04 -5.139084e-05 -1.606879e-05
#>                          C25           C26           C27           C28
#> brahma         -2.168918e-04  5.311528e-06 -1.087851e-04 -1.129037e-04
#> caney           1.277543e-04  1.380997e-04  2.220567e-04  3.358853e-04
#> chimay         -1.984526e-05 -2.262631e-04  9.411602e-05 -1.216533e-04
#> corona         -1.099958e-04 -2.321293e-04  9.791988e-05 -3.093851e-04
#> deusventrue    -3.818007e-04 -1.202012e-04 -8.165642e-05 -2.065725e-04
#> duvel           6.570113e-05  1.163622e-05 -1.505426e-04 -3.998066e-04
#> franziskaner    1.056867e-04  2.160659e-04  9.652675e-05 -1.456826e-04
#>                          C29           C30           C31           C32
#> brahma          5.059658e-05 -9.818126e-06  5.123235e-06 -1.428490e-04
#> caney           1.965508e-05  1.361109e-04  1.243089e-04 -1.336020e-04
#> chimay         -9.070337e-05 -2.449328e-04 -6.708875e-05  4.022984e-05
#> corona          1.431245e-04 -2.173931e-04 -2.259074e-04 -2.787437e-04
#> deusventrue    -1.783859e-04 -4.874278e-06 -3.212311e-04 -1.202210e-04
#> duvel          -1.385701e-04  6.771122e-05  1.314672e-04  1.274480e-04
#> franziskaner   -3.544520e-05 -1.570021e-04  4.143696e-05 -3.250007e-05
#>                       D1          D2         D3           D4         D5
#> brahma         0.2937120 -0.04602927 0.05240292 -0.035768593 0.03999516
#> caney          0.3046235 -0.07069129 0.05062805 -0.011400633 0.04383297
#> chimay         0.4156841 -0.09356117 0.04692603 -0.019249436 0.03965332
#> corona         0.2745921 -0.05755121 0.05150878 -0.011252954 0.03689351
#> deusventrue    0.3149661 -0.11964363 0.05529900  0.007135060 0.03861865
#> duvel          0.4496172 -0.09170033 0.05080071 -0.024018306 0.03036868
#> franziskaner   0.3002734 -0.05637154 0.04411627 -0.030282997 0.03014850
#>                           D6            D7            D8            D9
#> brahma          1.156917e-02  1.544573e-02  0.0013278090  0.0017860170
#> caney           4.485691e-03  9.597789e-03  0.0016029758  0.0082459711
#> chimay          1.731744e-02  1.044446e-02  0.0104067305  0.0026404547
#> corona         -4.664986e-03  9.465002e-03  0.0001838085  0.0120730321
#> deusventrue    -8.770510e-04  1.356919e-02  0.0076757796  0.0034745699
#> duvel           1.486796e-02  7.642213e-03  0.0154756340  0.0060020330
#> franziskaner    1.266218e-03  1.748451e-02  0.0047455611  0.0067787344
#>                         D10           D11          D12           D13
#> brahma          0.006789017  0.0048196384  0.007143143  1.926114e-03
#> caney           0.011464283  0.0039468981  0.003663644 -7.807175e-04
#> chimay          0.012590865  0.0018154608  0.002625857 -8.084753e-04
#> corona          0.008346831  0.0027556006  0.002567053  5.432929e-04
#> deusventrue     0.006022626  0.0007085549  0.005811255  2.407200e-04
#> duvel           0.014318246  0.0021405582 -0.001168779 -3.377569e-03
#> franziskaner    0.007215285  0.0030768626  0.006498196  8.028161e-04
#>                          D14           D15           D16           D17
#> brahma          0.0031275641  4.157794e-04  0.0019555535  9.890925e-04
#> caney           0.0037093445  1.744155e-03  0.0009874891 -4.753255e-04
#> chimay          0.0009166164  2.220375e-03 -0.0019270610  2.484912e-04
#> corona          0.0058977403  1.606741e-03  0.0024960621 -7.639544e-04
#> deusventrue     0.0033919410  1.559680e-04  0.0015420583  9.082901e-04
#> duvel          -0.0029313189  2.502233e-03  0.0006963735  1.208427e-03
#> franziskaner    0.0041872022  9.225601e-04  0.0024966790  6.722121e-04
#>                          D18           D19           D20           D21
#> brahma          6.293376e-04  3.896704e-04 -3.229345e-04  3.831138e-06
#> caney          -7.645439e-04 -6.085621e-04 -3.245536e-04 -4.503156e-04
#> chimay         -1.130345e-04  7.926418e-04 -1.780914e-04 -4.887150e-04
#> corona          1.024423e-03  8.842135e-05  4.313700e-04 -4.718718e-04
#> deusventrue    -3.583324e-04  1.286906e-03  6.233120e-05  1.544907e-03
#> duvel          -5.922990e-04 -1.333335e-04  4.815197e-04 -1.758102e-04
#> franziskaner    6.100396e-04  7.764117e-04  5.491195e-05  3.763780e-04
#>                          D22           D23           D24           D25
#> brahma         -6.023925e-04 -3.704707e-04 -8.677815e-04 -7.248682e-04
#> caney          -1.173466e-03 -3.863544e-04 -1.351268e-03 -1.773197e-04
#> chimay         -6.522348e-04 -7.978252e-04 -5.495898e-06 -1.800777e-03
#> corona         -8.788427e-04 -4.830025e-05 -5.870134e-04  7.484502e-04
#> deusventrue     1.415082e-05  7.733248e-04  3.116257e-05 -6.374096e-04
#> duvel          -2.941583e-04 -1.124154e-03 -1.118860e-03 -1.186668e-03
#> franziskaner   -3.051374e-04  2.712854e-04 -3.334307e-04  6.527762e-05
#>                          D26           D27           D28           D29
#> brahma         -6.395018e-04 -4.053179e-04 -2.320532e-04 -3.997015e-04
#> caney          -2.359438e-04 -2.059137e-04 -5.363116e-05 -5.687334e-04
#> chimay         -1.367853e-03 -1.301151e-03 -1.233134e-03 -5.067907e-04
#> corona         -5.900023e-04  4.867919e-04 -1.429390e-04  4.416367e-04
#> deusventrue    -5.512986e-04 -9.265221e-04 -4.809160e-04 -5.674634e-04
#> duvel          -6.434708e-04 -3.994538e-04 -2.003070e-04 -2.293221e-04
#> franziskaner   -1.310487e-04  2.683871e-04  8.363468e-06  1.249318e-04
#>                          D30           D31           D32
#> brahma         -3.902515e-04 -4.657627e-04 -4.591314e-04
#> caney          -9.850469e-05  3.166971e-05  1.112523e-04
#> chimay         -8.560520e-04 -4.323199e-04  9.946603e-05
#> corona          2.091023e-04 -7.157752e-05 -1.300100e-04
#> deusventrue     1.040810e-04 -8.204937e-04 -5.065759e-04
#> duvel          -3.433012e-05  3.583444e-04  3.946036e-04
#> franziskaner   -9.090693e-05 -3.654028e-04 -3.017993e-04
#>  [ reached getOption("max.print") -- omitted 33 rows ]

# if you want, say the first 8 harmonics but not the first one
retain <- coeff_sel(retain=8, drop=1, nb.h=32, cph=4)
head(coe[, retain])

#>                      A2         A3         A4          A5          A6
#> brahma      0.006766531 0.09348184 0.01374288 0.023818571 0.008592275
#> caney       0.007368844 0.09542847 0.01449765 0.022207326 0.006665955
#> chimay      0.016535135 0.07571705 0.02736989 0.009876978 0.014306122
#> corona      0.007486279 0.09616977 0.01220177 0.022887407 0.007281456
#> deusventrue 0.019627538 0.09281761 0.02230953 0.015264921 0.014857953
#> duvel       0.022177771 0.06895598 0.03354560 0.009316683 0.021131541
#>                       A7          A8            B2            B3            B4
#> brahma       0.003183571 0.005158502 -0.0001900652  3.306231e-04 -0.0005191749
#> caney        0.003552709 0.007010166  0.0005012013 -3.851293e-04  0.0003333918
#> chimay      -0.004741288 0.007814037  0.0001843629  4.196107e-04  0.0003227901
#> corona       0.005504589 0.007852411 -0.0003586724  1.711055e-05 -0.0005501057
#> deusventrue  0.002521451 0.011391904  0.0001774985 -8.326845e-05 -0.0014033732
#> duvel       -0.001687129 0.011025502 -0.0004198782  7.447638e-05 -0.0006627095
#>                        B5            B6            B7            B8
#> brahma       1.067446e-04 -9.218905e-05  5.604686e-07  6.268503e-06
#> caney       -2.899903e-04  7.350207e-05 -4.952054e-04  8.536695e-05
#> chimay      -2.906714e-05  5.573360e-04  1.059517e-04  6.209192e-04
#> corona      -1.907425e-04 -4.256287e-04 -2.147013e-04 -1.931107e-04
#> deusventrue -3.240180e-04 -9.330047e-04  6.515692e-04 -8.354423e-04
#> duvel        6.107940e-05 -4.746985e-04  2.450959e-04 -1.676532e-04
#>                        C2            C3            C4            C5
#> brahma      -0.0016375711 -3.936895e-03  0.0054080962 -1.259407e-03
#> caney        0.0012398277 -2.845651e-04  0.0003757825 -4.017802e-05
#> chimay      -0.0037576080 -1.797357e-03 -0.0021279238 -4.663387e-04
#> corona      -0.0016528639  1.573302e-03  0.0004897281  1.867708e-04
#> deusventrue  0.0015527752  7.706329e-04 -0.0014164476  1.463377e-03
#> duvel        0.0002872128 -5.422392e-06 -0.0007785717  3.178998e-04
#>                        C6            C7            C8          D2         D3
#> brahma      -3.994402e-03  3.268582e-03  0.0004792269 -0.04602927 0.05240292
#> caney        4.699805e-04  2.518166e-04  0.0006300072 -0.07069129 0.05062805
#> chimay      -7.424827e-05 -8.096453e-05  0.0009946667 -0.09356117 0.04692603
#> corona       6.888736e-04  3.145355e-04  0.0005189042 -0.05755121 0.05150878
#> deusventrue  9.055123e-04 -2.834923e-04 -0.0014439096 -0.11964363 0.05529900
#> duvel        2.219253e-04  5.438377e-04 -0.0000035189 -0.09170033 0.05080071
#>                      D4         D5           D6          D7           D8
#> brahma      -0.03576859 0.03999516  0.011569168 0.015445729 0.0013278090
#> caney       -0.01140063 0.04383297  0.004485691 0.009597789 0.0016029758
#> chimay      -0.01924944 0.03965332  0.017317438 0.010444458 0.0104067305
#> corona      -0.01125295 0.03689351 -0.004664986 0.009465002 0.0001838085
#> deusventrue  0.00713506 0.03861865 -0.000877051 0.013569192 0.0076757796
#> duvel       -0.02401831 0.03036868  0.014867959 0.007642213 0.0154756340