Data: Outline coordinates of 360 molars

Courtesy of Julien Corny and Florent Detroit.

Format

A Out object containing 79 equilinearly spaced (x; y) coordinates for 360 crown outlines, of modern human molars, along with their type ($type) - 90 first upper molars (UM1), 90 second upper molars (UM2), 90 first lower molars (LM1), 90 second lower molars (LM2) - and the individual (ind) they come from (the data of the 360 molars are taken from 180 individuals).

Source

Corny, J., & Detroit, F. (2014). Technical Note: Anatomic identification of isolated modern human molars: testing Procrustes aligned outlines as a standardization procedure for elliptic fourier analysis. American Journal of Physical Anthropology, 153(2), 314-22. doi:10.1002/ajpa.22428 see https://onlinelibrary.wiley.com/doi/abs/10.1002/ajpa.22428

See also