Data: Outline coordinates of North Sea fishes

Data: Outline coordinates of North Sea fishes

Format

A Out object containing the outlines coordinates for 218 fishes from the North Sea along with taxonomical cofactors.

Source

Caillon F, Frelat R, Mollmann C, Bonhomme V (submitted)

See also

Other datasets: apodemus, bot, chaff, charring, flower, hearts, molars, mosquito, mouse, oak, olea, shapes, trilo, wings