Data: Outline coordinates of cephalic outlines of trilobite

Data: Outline coordinates of cephalic outlines of trilobite

Format

A Out object 64 coordinates of 50 cephalic outlines from different ontogenetic stages of trilobite.

Source

Arranged from: http://folk.uio.no (used to be in ohammer website but seems to be deprecated now). The original data included 51 outlines and 5 ontogenetic stages, but one of them has just a single outline thas has been removed.

See also