Defines sliding landmarks matrix

Defines sliding landmarks matrix

def_slidings(Coo, slidings)

Arguments

Coo
an Ldk object

slidings
a matrix, a numeric or a list of numeric. See Details

Details

$slidings in Ldk must be a 'valid' matrix: containing ids of coordinates, none of them being lower than 1 and higher the number of coordinates in $coo. slidings matrix contains 3 columns (before, slide, after). It is inspired by geomorph and should be compatible with it.

This matrix can be passed directly if the slidings argument is a matrix. Of course, it is strictly equivalent to Ldk$slidings <- slidings. slidings can also be passed as "partition(s)", when sliding landmarks identified by their ids (which are a row number) are consecutive in the $coo.

A single partition can be passed either as a numeric (eg 4:12), if points 5 to 11 must be considered as sliding landmarks (4 and 12 being fixed); or as a list of numeric.

See examples below.

See also

Other ldk/slidings methods: add_ldk, def_ldk, get_ldk, get_slidings, rearrange_ldk, slidings_scheme

Examples

#waiting for a sliding dataset...