Import files creates by StereoMorph into Momocs

Helps to read .txt files created by StereoMorph into (x; y) coordinates or Momocs objects. Can be applied to 'curves' or 'ldk' text files.

Usage

import_StereoMorph_curve1(path)

import_StereoMorph_curve(path, names)

import_StereoMorph_ldk1(path)

import_StereoMorph_ldk(path, names)

Arguments

path

toward a single file or a folder containing .txt files produced by StereoMorph

names

to feed lf_structure

Value

a list of class Coo

Details

*1 functions import a single .txt file. Their counterpart (no '1') work when path indicates the folder, i.e. 'curves' or 'ldk'. They then return a list of Opn or Ldk objects, respectively. Please do not hesitate to contact me should you have a particular case or need something.

Note

Note this function will be deprecated from Momocs when Momacs and Momit will be fully operationnal.

See also