grindr layers for multivariate plots

Useful layers for building custom mutivariate plots using the cheapbabi approach. See examples.

layer_frame(x, center_origin = TRUE, zoom = 0.9)

layer_axes(x, col = "#999999", lwd = 1/2, ...)

layer_ticks(x, col = "#333333", cex = 3/4, lwd = 3/4, ...)

layer_grid(x, col = "#999999", lty = 3, grid = 3, ...)

layer_box(x, border = "#e5e5e5", ...)

layer_fullframe(x, ...)

layer_points(x, pch = 20, cex = 4/log1p(nrow(x$xy)), transp = 0, ...)

layer_ellipses(x, conf = 0.5, lwd = 1, alpha = 0, ...)

layer_ellipsesfilled(x, conf = 0.5, lwd = 1, alpha = 0, ...)

layer_ellipsesaxes(x, conf = 0.5, lwd = 1, alpha = 0, ...)

layer_chull(x, ...)

layer_chullfilled(x, alpha = 0.8, ...)

layer_stars(x, alpha = 0.5, ...)

layer_delaunay(x, ...)

layer_density(
  x,
  levels_density = 20,
  levels_contour = 4,
  alpha = 1/3,
  n = 200,
  density = TRUE,
  contour = TRUE
)

layer_labelpoints(
  x,
  col = par("fg"),
  cex = 2/3,
  font = 1,
  abbreviate = FALSE,
  ...
)

layer_labelgroups(
  x,
  col = par("fg"),
  cex = 3/4,
  font = 2,
  rect = TRUE,
  alpha = 1/4,
  abbreviate = FALSE,
  ...
)

layer_rug(x, size = 1/200, ...)

layer_histogram_2(x, freq = FALSE, breaks, split = FALSE, transp = 0)

layer_density_2(x, bw, split = FALSE, rug = TRUE, transp = 0)

layer_title(x, title = "", cex = 3/4, ...)

layer_axesnames(x, cex = 3/4, name = "Axis", ...)

layer_eigen(x, nb_max = 5, cex = 1/2, ...)

layer_axesvar(x, cex = 3/4, ...)

layer_legend(x, probs = seq(0, 1, 0.25), cex = 3/4, ...)

Arguments

x
a list, typically returned by plot_PCA

center_origin
logical whether to center the origin (default TRUE)

zoom
numeric to change the zoom (default 0.9)

col
color (hexadecimal) to use for drawing components

lwd
linewidth for drawing components

...
additional options to feed core functions for each layer

cex
to use for drawing components

lty
linetype for drawing components

grid
numeric number of grid to draw

border
color (hexadecimal) to use to draw border

pch
to use for drawing components

transp
transparency to use (min: 0 defaut:0 max:1)

conf
numeric between 0 and 1 for confidence ellipses

alpha
numeric between 0 and 1 for the transparency of components

levels_density
numeric number of levels to use to feed MASS::kde2d

levels_contour
numeric number of levels to use to feed graphics::contour

n
numeric number of grid points to feed MASS::kde2d

density
logical whether to draw density estimate

contour
logical whether to draw contour lines

font
to feed text

abbreviate
logical whether to abbreviate names

rect
logical whether to draw a rectangle below names

size
numeric as a fraction of graphical window (default: 1/200)

freq
logicalto feed[hist] (default:FALSE`)

breaks
to feed hist (default: calculated on the pooled values)

split
logical whether to split the two distributions into two plots

bw
to feed density (default: stats::bw.nrd0)

rug
logical whether to add rug (default: TRUE)

title
to add to the plot (default "")

name
to use on axes (default "Axis")

nb_max
numeric number of eigen values to display (default 5)

probs
numeric sequence to feed stats::quantile and to indicate where to draw ticks and legend labels

See also