List available methods for any object and other classes for a method

Given any object returns all available methods for its class, or given a method list all other supported classes

list_methods_for_this_class(x)

list_classes_for_this_method(x)

Arguments

x
any object

Examples

#methods for data.frame
list_methods_for_this_class(iris)

#> * 'data.frame' class:
#>   - [.data.frame
#>   - [[.data.frame
#>   - [[<-,data.frame,ANY,ANY-method
#>   - [[<-.data.frame
#>   - [<-,data.frame,ANY,ANY,ANY-method
#>   - [<-.data.frame
#>   - $.data.frame
#>   - $<-,data.frame-method
#>   - $<-.data.frame
#>   - addAttrToGeom,SpatialGrid,data.frame-method
#>   - addAttrToGeom,SpatialLines,data.frame-method
#>   - addAttrToGeom,SpatialMultiPoints,data.frame-method
#>   - addAttrToGeom,SpatialPixels,data.frame-method
#>   - addAttrToGeom,SpatialPoints,data.frame-method
#>   - addAttrToGeom,SpatialPolygons,data.frame-method
#>   - aggregate.data.frame
#>   - anti_join.data.frame
#>   - anyDuplicated.data.frame
#>   - arrange_.data.frame
#>   - arrange.data.frame
#>   - as_data_frame.data.frame
#>   - as_tibble.data.frame
#>   - as.data.frame.data.frame
#>   - as.list.data.frame
#>   - as.matrix.data.frame
#>   - as.tbl_cube.data.frame
#>   - as.tbl.data.frame
#>   - by.data.frame
#>   - cbind.data.frame
#>   - coerce,abIndex,vector-method
#>   - coerce,ANY,list-method
#>   - coerce,data.frame,gpc.poly-method
#>   - coerce,dgCMatrix,vector-method
#>   - coerce,dsCMatrix,vector-method
#>   - coerce,GridTopology,data.frame-method
#>   - coerce,list,indMatrix-method
#>   - coerce,Matrix,vector-method
#>   - coerce,oldClass,S3-method
#>   - coerce,sparseVector,vector-method
#>   - coerce,SpatialGrid,data.frame-method
#>   - coerce,SpatialGridDataFrame,data.frame-method
#>   - coerce,SpatialLinesDataFrame,data.frame-method
#>   - coerce,SpatialMultiPoints,data.frame-method
#>   - coerce,SpatialMultiPointsDataFrame,data.frame-method
#>   - coerce,SpatialPixels,data.frame-method
#>   - coerce,SpatialPixelsDataFrame,data.frame-method
#>   - coerce,SpatialPoints,data.frame-method
#>   - coerce,SpatialPointsDataFrame,data.frame-method
#>   - coerce,SpatialPolygonsDataFrame,data.frame-method
#>   - coerce,SpatialRingsDataFrame,data.frame-method
#>   - collapse.data.frame
#>   - collect.data.frame
#>   - compute.data.frame
#>   - coordinates,data.frame-method
#>   - coordinates<-,data.frame-method
#>   - dim.data.frame
#>   - dimnames.data.frame
#>   - dimnames<-.data.frame
#>   - distinct_.data.frame
#>   - distinct.data.frame
#>   - do_.data.frame
#>   - do.data.frame
#>   - droplevels.data.frame
#>   - duplicated.data.frame
#>   - edit.data.frame
#>   - filter_.data.frame
#>   - filter.data.frame
#>   - format.data.frame
#>   - formula.data.frame
#>   - fortify.data.frame
#>   - full_join.data.frame
#>   - geometry<-,data.frame,Spatial-method
#>   - ggplot.data.frame
#>   - glimpse.data.frame
#>   - gridded<-,data.frame,character-method
#>   - gridded<-,data.frame,formula-method
#>   - gridded<-,data.frame,GridTopology-method
#>   - gridded<-,SpatialPoints,list-method
#>   - gridded<-,SpatialPointsDataFrame,list-method
#>   - group_by_.data.frame
#>   - group_by.data.frame
#>   - group_indices_.data.frame
#>   - group_indices.data.frame
#>   - group_size.data.frame
#>   - groups.data.frame
#>   - head.data.frame
#>   - initialize,oldClass-method
#>   - inner_join.data.frame
#>   - intersect.data.frame
#>   - is.na.data.frame
#>   - lda.data.frame
#>   - left_join.data.frame
#>   - Math.data.frame
#>   - merge,Spatial,data.frame-method
#>   - merge.data.frame
#>   - mutate_.data.frame
#>   - mutate.data.frame
#>   - n_groups.data.frame
#>   - na.exclude.data.frame
#>   - na.omit.data.frame
#>   - Ops,nonStructure,vector-method
#>   - Ops,structure,vector-method
#>   - Ops,vector,nonStructure-method
#>   - Ops,vector,structure-method
#>   - Ops.data.frame
#>   - plot.data.frame
#>   - polygons<-,data.frame,SpatialPolygons-method
#>   - print.data.frame
#>   - prompt.data.frame
#>   - pull.data.frame
#>   - rbind.data.frame
#>   - rename_.data.frame
#>   - rename.data.frame
#>   - right_join.data.frame
#>   - row.names.data.frame
#>   - row.names<-.data.frame
#>   - rowsum.data.frame
#>   - same_src.data.frame
#>   - sample_frac.data.frame
#>   - sample_n.data.frame
#>   - select_.data.frame
#>   - select.data.frame
#>   - semi_join.data.frame
#>   - setdiff.data.frame
#>   - setequal.data.frame
#>   - show,oldClass-method
#>   - slice_.data.frame
#>   - slice.data.frame
#>   - slotsFromS3,data.frame-method
#>   - split.data.frame
#>   - split<-.data.frame
#>   - sppanel,list-method
#>   - stack.data.frame
#>   - str.data.frame
#>   - subset.data.frame
#>   - summarise_.data.frame
#>   - summarise.data.frame
#>   - summary.data.frame
#>   - Summary.data.frame
#>   - t.data.frame
#>   - tail.data.frame
#>   - tbl_vars.data.frame
#>   - transform.data.frame
#>   - type_sum.data.frame
#>   - ungroup.data.frame
#>   - union_all.data.frame
#>   - union.data.frame
#>   - unique.data.frame
#>   - unstack.data.frame
#>   - within.data.frame

#methods for Coo objects
list_methods_for_this_class(bot)

#> * 'Out' class:
#>   - add_ldk.Out
#>   - calibrate_deviations.Out
#>   - calibrate_harmonicpower.Out
#>   - calibrate_reconstructions.Out
#>   - combine.Out
#>   - coo_bookstein.Out
#>   - coo_down.Out
#>   - coo_left.Out
#>   - coo_right.Out
#>   - coo_sample_prop.Out
#>   - coo_sample.Out
#>   - coo_slice.Out
#>   - coo_up.Out
#>   - d.Out
#>   - def_ldk_angle.Out
#>   - def_ldk_direction.Out
#>   - def_ldk.Out
#>   - efourier.Out
#>   - fgProcrustes.Out
#>   - get_ldk.Out
#>   - panel.Out
#>   - print.Out
#>   - rearrange_ldk.Out
#>   - rfourier.Out
#>   - sfourier.Out
#>   - tfourier.Out
#> * 'Coo' class:
#>   - [.Coo
#>   - [<-.Coo
#>   - $.Coo
#>   - arrange.Coo
#>   - as_df.Coo
#>   - chop.Coo
#>   - coo_align.Coo
#>   - coo_aligncalliper.Coo
#>   - coo_alignminradius.Coo
#>   - coo_alignxax.Coo
#>   - coo_angle_edges.Coo
#>   - coo_angle_tangent.Coo
#>   - coo_area.Coo
#>   - coo_baseline.Coo
#>   - coo_boundingbox.Coo
#>   - coo_calliper.Coo
#>   - coo_centdist.Coo
#>   - coo_center.Coo
#>   - coo_centpos.Coo
#>   - coo_centsize.Coo
#>   - coo_check.Coo
#>   - coo_chull_onion.Coo
#>   - coo_chull.Coo
#>   - coo_circularity.Coo
#>   - coo_circularityharalick.Coo
#>   - coo_circularitynorm.Coo
#>   - coo_close.Coo
#>   - coo_convexity.Coo
#>   - coo_down.Coo
#>   - coo_dxy.Coo
#>   - coo_eccentricityboundingbox.Coo
#>   - coo_eccentricityeigen.Coo
#>   - coo_elongation.Coo
#>   - coo_extract.Coo
#>   - coo_flipx.Coo
#>   - coo_flipy.Coo
#>   - coo_force2close.Coo
#>   - coo_interpolate.Coo
#>   - coo_intersect_angle.Coo
#>   - coo_intersect_direction.Coo
#>   - coo_intersect_segment.Coo
#>   - coo_jitter.Coo
#>   - coo_left.Coo
#>   - coo_length.Coo
#>   - coo_lw.Coo
#>   - coo_nb.Coo
#>   - coo_perim.Coo
#>   - coo_perimcum.Coo
#>   - coo_perimpts.Coo
#>   - coo_range_enlarge.Coo
#>   - coo_range.Coo
#>   - coo_rectangularity.Coo
#>   - coo_rectilinearity.Coo
#>   - coo_rev.Coo
#>   - coo_right.Coo
#>   - coo_rotate.Coo
#>   - coo_rotatecenter.Coo
#>   - coo_samplerr.Coo
#>   - coo_scale.Coo
#>   - coo_scalex.Coo
#>   - coo_scaley.Coo
#>   - coo_shearx.Coo
#>   - coo_sheary.Coo
#>   - coo_slide.Coo
#>   - coo_slidedirection.Coo
#>   - coo_slidegap.Coo
#>   - coo_smooth.Coo
#>   - coo_solidity.Coo
#>   - coo_template.Coo
#>   - coo_trans.Coo
#>   - coo_trim.Coo
#>   - coo_trimbottom.Coo
#>   - coo_trimtop.Coo
#>   - coo_unclose.Coo
#>   - coo_up.Coo
#>   - coo_width.Coo
#>   - dfourier.Coo
#>   - dim.Coo
#>   - filter.Coo
#>   - is_clockwise.Coo
#>   - is_closed.Coo
#>   - is_equallyspacedradii.Coo
#>   - Ldk.Coo
#>   - length.Coo
#>   - measure.Coo
#>   - mutate.Coo
#>   - names.Coo
#>   - names<-.Coo
#>   - Opn.Coo
#>   - Out.Coo
#>   - plot.Coo
#>   - rename.Coo
#>   - sample_frac.Coo
#>   - sample_n.Coo
#>   - select.Coo
#>   - slice.Coo
#>   - stack.Coo
#>   - str.Coo
#>   - subset.Coo
#>   - table.Coo
#>   - validate.Coo

#methods for Coe objects
list_methods_for_this_class(bot %>% efourier)

#> 'nb.h' not provided and set to 10 (99% harmonic power)

#> * 'OutCoe' class:
#>   - boxplot.OutCoe
#>   - combine.OutCoe
#>   - hcontrib.OutCoe
#>   - hist.OutCoe
#>   - LDA.OutCoe
#>   - MANOVA.OutCoe
#>   - mshapes.OutCoe
#>   - panel.OutCoe
#>   - PCA.OutCoe
#>   - print.OutCoe
#>   - rm_asym.OutCoe
#>   - rm_sym.OutCoe
#>   - symmetry.OutCoe
#> * 'Coe' class:
#>   - [.Coe
#>   - [<-.Coe
#>   - $.Coe
#>   - arrange.Coe
#>   - as_df.Coe
#>   - breed.Coe
#>   - chop.Coe
#>   - classify.Coe
#>   - CLUST.Coe
#>   - dim.Coe
#>   - dissolve.Coe
#>   - export.Coe
#>   - filter.Coe
#>   - get_pairs.Coe
#>   - LDA.Coe
#>   - length.Coe
#>   - mutate.Coe
#>   - names.Coe
#>   - names<-.Coe
#>   - perm.Coe
#>   - reLDA.Coe
#>   - rename.Coe
#>   - sample_frac.Coe
#>   - sample_n.Coe
#>   - select.Coe
#>   - slice.Coe
#>   - str.Coe
#>   - subset.Coe
#>   - table.Coe
#>   - which_out.Coe

#methods for PCA objects
list_methods_for_this_class(bot %>% efourier %>% PCA)

#> 'nb.h' not provided and set to 10 (99% harmonic power)

#> * 'PCA' class:
#>   - $.PCA
#>   - arrange.PCA
#>   - as_df.PCA
#>   - boxplot.PCA
#>   - CLUST.PCA
#>   - export.PCA
#>   - filter.PCA
#>   - flip_PCaxes.PCA
#>   - get_pairs.PCA
#>   - KMEANS.PCA
#>   - LDA.PCA
#>   - MANOVA_PW.PCA
#>   - MANOVA.PCA
#>   - mshapes.PCA
#>   - mutate.PCA
#>   - PCcontrib.PCA
#>   - plot.PCA
#>   - print.PCA
#>   - reLDA.PCA
#>   - rename.PCA
#>   - rePCA.PCA
#>   - scree.PCA
#>   - select.PCA
#>   - slice.PCA
#>   - subset.PCA
#>   - table.PCA
#> * 'prcomp' class:
#>   - biplot.prcomp
#>   - plot.prcomp
#>   - predict.prcomp
#>   - print.prcomp
#>   - summary.prcomp

# classes supported for
list_classes_for_this_method(PCAAA) # nothing

#> NULL

list_classes_for_this_method(PCA)

#>   - PCA.default
#>   - PCA.LdkCoe
#>   - PCA.OpnCoe
#>   - PCA.OutCoe
#>   - PCA.TraCoe

list_classes_for_this_method(sum) # nothing

#> NULL