NMDS plot unsing grindr layers

Quickly vizualise MDS and NMDS objects and build customs plots using the layers. See examples.

plot_NMDS(
  x,
  f = NULL,
  axes = c(1, 2),
  points = TRUE,
  points_transp = 1/4,
  chull = TRUE,
  chullfilled = FALSE,
  labelgroups = FALSE,
  legend = TRUE,
  title = "",
  box = TRUE,
  axesnames = TRUE,
  palette = pal_qual
)

plot_MDS(
  x,
  f = NULL,
  axes = c(1, 2),
  points = TRUE,
  points_transp = 1/4,
  chull = TRUE,
  chullfilled = FALSE,
  labelgroups = FALSE,
  legend = TRUE,
  title = "",
  box = TRUE,
  axesnames = TRUE,
  palette = pal_qual
)

Arguments

x
the result of MDS or NMDS

f
factor specification to feed fac_dispatcher

axes
numeric of length two to select PCs to use (c(1, 2) by default)

points
logical whether to draw this with layer_points

points_transp
numeric to feed layer_points (default:0.25)

chull
logical whether to draw this with layer_chull

chullfilled
logical whether to draw this with layer_chullfilled

labelgroups
logical whether to draw this with layer_labelgroups

legend
logical whether to draw this with layer_legend

title
character if specified, fee layer_title (default to "")

box
logical whether to draw this using layer_box

axesnames
logical whether to draw this using layer_axesnames

palette
color palette to use col_summer by default

See also

Examples

### First prepare an NMDS object
x <- bot %>% efourier %>% NMDS
#> 'norm=TRUE' is used and this may be troublesome. See ?efourier
#> 'nb.h' set to 10 (99% harmonic power)
#> 'comm' has negative data: 'autotransform', 'noshare' and 'wascores' set to FALSE
#> Warning: results may be meaningless because data have negative entries in method “bray”
#> Run 0 stress 0.07227125 
#> Run 1 stress 0.07227125 
#> ... Procrustes: rmse 5.189443e-06  max resid 1.723229e-05 
#> ... Similar to previous best
#> Run 2 stress 0.07227125 
#> ... New best solution
#> ... Procrustes: rmse 4.652192e-06  max resid 1.792204e-05 
#> ... Similar to previous best
#> Run 3 stress 0.07227125 
#> ... Procrustes: rmse 2.082389e-06  max resid 7.278089e-06 
#> ... Similar to previous best
#> Run 4 stress 0.07227125 
#> ... Procrustes: rmse 7.917109e-06  max resid 3.563239e-05 
#> ... Similar to previous best
#> Run 5 stress 0.07227125 
#> ... Procrustes: rmse 1.11719e-05  max resid 4.420689e-05 
#> ... Similar to previous best
#> Run 6 stress 0.07227125 
#> ... Procrustes: rmse 7.0436e-06  max resid 2.832968e-05 
#> ... Similar to previous best
#> Run 7 stress 0.1659489 
#> Run 8 stress 0.07227125 
#> ... Procrustes: rmse 5.333644e-06  max resid 2.688398e-05 
#> ... Similar to previous best
#> Run 9 stress 0.1717542 
#> Run 10 stress 0.07227125 
#> ... Procrustes: rmse 5.169194e-06  max resid 2.057508e-05 
#> ... Similar to previous best
#> Run 11 stress 0.07227125 
#> ... Procrustes: rmse 5.793606e-06  max resid 2.240695e-05 
#> ... Similar to previous best
#> Run 12 stress 0.07227125 
#> ... Procrustes: rmse 2.436606e-06  max resid 1.206057e-05 
#> ... Similar to previous best
#> Run 13 stress 0.07227125 
#> ... Procrustes: rmse 5.674202e-06  max resid 2.206627e-05 
#> ... Similar to previous best
#> Run 14 stress 0.07227125 
#> ... Procrustes: rmse 6.183384e-06  max resid 2.464175e-05 
#> ... Similar to previous best
#> Run 15 stress 0.07227126 
#> ... Procrustes: rmse 1.838659e-05  max resid 7.303832e-05 
#> ... Similar to previous best
#> Run 16 stress 0.07227125 
#> ... Procrustes: rmse 6.291716e-06  max resid 3.235569e-05 
#> ... Similar to previous best
#> Run 17 stress 0.07227125 
#> ... Procrustes: rmse 8.448117e-06  max resid 3.641249e-05 
#> ... Similar to previous best
#> Run 18 stress 0.07227125 
#> ... Procrustes: rmse 4.566722e-06  max resid 1.738098e-05 
#> ... Similar to previous best
#> Run 19 stress 0.07227125 
#> ... Procrustes: rmse 1.744469e-06  max resid 5.633608e-06 
#> ... Similar to previous best
#> Run 20 stress 0.07227125 
#> ... Procrustes: rmse 4.729808e-06  max resid 1.898942e-05 
#> ... Similar to previous best
#> *** Solution reached

plot_NMDS(x)
#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet
plot_NMDS(x, ~type) %>% layer_stars() %>% layer_labelpoints()
#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet

### Same on MDS object
x <- bot %>% efourier %>% MDS
#> 'norm=TRUE' is used and this may be troublesome. See ?efourier
#> 'nb.h' set to 10 (99% harmonic power)

plot_MDS(x)
#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet
plot_MDS(x, ~type) %>% layer_stars() %>% layer_labelpoints()
#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet