Builds an Ldk object

In Momocs, Ldk classes objects are lists of configurations of landmarks, with optionnal components, on which generic methods such as plotting methods (e.g. stack) and specific methods (e.g. fgProcrustes). Ldk objects are primarily Coo objects. In a sense, morphometrics methods on Ldk objects preserves (x, y) coordinates and LdkCoe are also Ldk objects.

Ldk(coo, links = NULL, slidings = NULL, fac = dplyr::data_frame())

Arguments

coo
a list of matrices of (x; y) coordinates, or an array, an Ldk object.

links
(optionnal) a 2-columns matrix of 'links' between landmarks, mainly for plotting

slidings
(optionnal) a 3-columns matrix defining (if any) sliding landmarks

fac
(optionnal) a data.frame of factors and/or numerics specifying the grouping structure

Value

an Ldk object

Details

All the shapes in x must have the same number of landmarks. If you are trying to make an Ldk object from an Out or an Opn object, try coo_sample beforehand to homogeneize the number of coordinates among shapes.

implementation of $slidings is inspired by geomorph

Examples

#Methods on Ldk
methods(class=Ldk)

#>  [1] combine         coo_bookstein   coo_slice       d              
#>  [5] def_links       def_slidings    fgProcrustes    fgsProcrustes  
#>  [9] get_ldk         get_slidings    mshapes         panel          
#> [13] pile            rearrange_ldk   slidings_scheme stack          
#> [17] which_out      
#> see '?methods' for accessing help and source code