Non metric multidimensional scaling

A wrapper around vegan::metaMDS.

NMDS(x, distance = "bray", k = 2, try = 20, trymax = 20, ...)

Arguments

x
any Coe object

distance
a dissiminarity index to feed vegan::vegdist (default: bray)

k
numeric number of dimensions to feed vegan::metaMDS (default: 2)

try
numeric minimum number of random starts to feed vegan::metaMDS (default: 20)

trymax
numeric minimum number of random starts to feed vegan::metaMDS (default: 20)

...
additional parameters to feed vegan::metaMDS

Value

what is returned by vegan::metaMDS plus $fac. And prepend NMDS class to it.

Details

For Details, see vegan::metaMDS

See also

Examples

x <- bot %>% efourier %>% NMDS

#> `norm=TRUE` is used and this may be troublesome. See ?efourier

#> 'nb.h' set to 10 (99% harmonic power)

#> Registered S3 method overwritten by 'vegan':
#>   method     from      
#>   rev.hclust dendextend

#> 'comm' has negative data: 'autotransform', 'noshare' and 'wascores' set to FALSE

#> Warning: results may be meaningless because data have negative entries in method “bray”

#> Run 0 stress 0.07227125 
#> Run 1 stress 0.07227132 
#> ... Procrustes: rmse 2.134479e-05  max resid 6.102186e-05 
#> ... Similar to previous best
#> Run 2 stress 0.1536611 
#> Run 3 stress 0.1729718 
#> Run 4 stress 0.07227128 
#> ... Procrustes: rmse 2.27572e-05  max resid 0.0001142424 
#> ... Similar to previous best
#> Run 5 stress 0.07227125 
#> ... Procrustes: rmse 8.797053e-06  max resid 3.993071e-05 
#> ... Similar to previous best
#> Run 6 stress 0.07227125 
#> ... Procrustes: rmse 5.388401e-06  max resid 2.000297e-05 
#> ... Similar to previous best
#> Run 7 stress 0.07227129 
#> ... Procrustes: rmse 4.13528e-05  max resid 0.0001738164 
#> ... Similar to previous best
#> Run 8 stress 0.1476538 
#> Run 9 stress 0.07227125 
#> ... New best solution
#> ... Procrustes: rmse 3.713068e-06  max resid 1.69194e-05 
#> ... Similar to previous best
#> Run 10 stress 0.07227125 
#> ... Procrustes: rmse 4.388959e-06  max resid 1.391617e-05 
#> ... Similar to previous best
#> Run 11 stress 0.07227131 
#> ... Procrustes: rmse 4.67699e-05  max resid 0.0001804122 
#> ... Similar to previous best
#> Run 12 stress 0.1647891 
#> Run 13 stress 0.07227126 
#> ... Procrustes: rmse 2.048337e-05  max resid 8.248206e-05 
#> ... Similar to previous best
#> Run 14 stress 0.07227126 
#> ... Procrustes: rmse 2.071888e-05  max resid 8.201421e-05 
#> ... Similar to previous best
#> Run 15 stress 0.1606942 
#> Run 16 stress 0.07227125 
#> ... Procrustes: rmse 8.159651e-06  max resid 3.149556e-05 
#> ... Similar to previous best
#> Run 17 stress 0.07227125 
#> ... Procrustes: rmse 1.742992e-06  max resid 7.302496e-06 
#> ... Similar to previous best
#> Run 18 stress 0.1776129 
#> Run 19 stress 0.07227125 
#> ... Procrustes: rmse 1.328271e-05  max resid 5.422646e-05 
#> ... Similar to previous best
#> Run 20 stress 0.07227125 
#> ... Procrustes: rmse 2.15242e-06  max resid 6.763911e-06 
#> ... Similar to previous best
#> *** Solution reached

# Shepard diagram # before a Momocs wrapper
# vegan::stressplot(x)