Non metric multidimensional scaling

A wrapper around vegan::metaMDS.

NMDS(x, distance = "bray", k = 2, try = 20, trymax = 20, ...)

Arguments

x
any Coe object

distance
a dissiminarity index to feed vegan::vegdist (default: bray)

k
numeric number of dimensions to feed vegan::metaMDS (default: 2)

try
numeric minimum number of random starts to feed vegan::metaMDS (default: 20)

trymax
numeric minimum number of random starts to feed vegan::metaMDS (default: 20)

...
additional parameters to feed vegan::metaMDS

Value

what is returned by vegan::metaMDS plus $fac. And prepend NMDS class to it.

Details

For Details, see vegan::metaMDS

See also

Examples

x <- bot %>% efourier %>% NMDS

#> `norm=TRUE` is used and this may be troublesome. See ?efourier

#> 'nb.h' set to 10 (99% harmonic power)

#> 'comm' has negative data: 'autotransform', 'noshare' and 'wascores' set to FALSE

#> Warning: results may be meaningless because data have negative entries in method “bray”

#> Run 0 stress 0.07227125 
#> Run 1 stress 0.07227126 
#> ... Procrustes: rmse 1.617493e-05  max resid 6.031545e-05 
#> ... Similar to previous best
#> Run 2 stress 0.07227125 
#> ... Procrustes: rmse 9.322816e-06  max resid 3.38855e-05 
#> ... Similar to previous best
#> Run 3 stress 0.07227127 
#> ... Procrustes: rmse 3.550582e-05  max resid 0.0001313918 
#> ... Similar to previous best
#> Run 4 stress 0.07227126 
#> ... Procrustes: rmse 2.426918e-05  max resid 9.204528e-05 
#> ... Similar to previous best
#> Run 5 stress 0.07227125 
#> ... Procrustes: rmse 4.726889e-06  max resid 1.445588e-05 
#> ... Similar to previous best
#> Run 6 stress 0.1684233 
#> Run 7 stress 0.07227126 
#> ... Procrustes: rmse 1.940177e-05  max resid 6.916372e-05 
#> ... Similar to previous best
#> Run 8 stress 0.0722713 
#> ... Procrustes: rmse 1.880684e-05  max resid 4.590807e-05 
#> ... Similar to previous best
#> Run 9 stress 0.07227125 
#> ... Procrustes: rmse 4.175003e-06  max resid 1.463801e-05 
#> ... Similar to previous best
#> Run 10 stress 0.07227125 
#> ... Procrustes: rmse 5.948362e-06  max resid 2.787275e-05 
#> ... Similar to previous best
#> Run 11 stress 0.07227129 
#> ... Procrustes: rmse 2.156687e-05  max resid 7.252355e-05 
#> ... Similar to previous best
#> Run 12 stress 0.1723591 
#> Run 13 stress 0.07227128 
#> ... Procrustes: rmse 3.718727e-05  max resid 0.0001561026 
#> ... Similar to previous best
#> Run 14 stress 0.07227125 
#> ... Procrustes: rmse 7.202872e-06  max resid 2.44823e-05 
#> ... Similar to previous best
#> Run 15 stress 0.1692495 
#> Run 16 stress 0.07227127 
#> ... Procrustes: rmse 3.023991e-05  max resid 0.000125352 
#> ... Similar to previous best
#> Run 17 stress 0.07227125 
#> ... Procrustes: rmse 5.73856e-06  max resid 1.841418e-05 
#> ... Similar to previous best
#> Run 18 stress 0.07227125 
#> ... Procrustes: rmse 3.792351e-06  max resid 2.003342e-05 
#> ... Similar to previous best
#> Run 19 stress 0.07227125 
#> ... Procrustes: rmse 8.642971e-06  max resid 2.135103e-05 
#> ... Similar to previous best
#> Run 20 stress 0.07227125 
#> ... Procrustes: rmse 8.377458e-06  max resid 2.278946e-05 
#> ... Similar to previous best
#> *** Solution reached

# Shepard diagram # before a Momocs wrapper
# vegan::stressplot(x)