Builds an OpnCoe object

In Momocs, OpnCoe classes objects are wrapping around lists of morphometric coefficients, along with other informations, on which generic methods such as plotting methods (e.g. boxplot) and specific methods can be applied. OpnCoe objects are primarily Coe objects.

Usage

OpnCoe(
  coe = matrix(),
  fac = dplyr::tibble(),
  method = character(),
  baseline1 = numeric(),
  baseline2 = numeric(),
  mod = list(),
  r2 = numeric()
)

Arguments

coe

matrix of morphometric coefficients

fac

(optionnal) a data.frame of factors, specifying the grouping structure

method

used to obtain these coefficients

baseline1

\((x; y)\) coordinates of the first baseline point

baseline2

\((x; y)\) coordinates of the second baseline point

mod

an R lm object, used to reconstruct shapes

r2

numeric, the r-squared from every model

Value

an OpnCoe object

See also

Examples

# all OpnCoe classes
methods(class='OpnCoe')
#> [1] MANOVA  MSHAPES PCA     boxplot combine print  
#> see '?methods' for accessing help and source code