Calculate classification metrics on a confusion matrix

In some cases, the class correctness or the proportion of correctly classified individuals is not enough, so here are more detailed metrics when working on classification.

classification_metrics(x)

Arguments

x
a table or an LDA object

Value

a list with the following components is returned:

accuracy the fraction of instances that are correctly classified

macro_prf data.frame containing precision (the fraction of correct predictions for a certain class); recall, the fraction of instances of a class that were correctly predicted; f1 the harmonic mean (or a weighted average) of precision and recall.

macro_avg, just the average of the three macro_prf indices

ova a list of one-vs-all confusion matrices for each class

ova_sum a single of all ova matrices

kappa measure of agreement between the predictions and the actual labels

See also

Examples

# some morphometrics on 'hearts'
hearts %>% fgProcrustes(tol=1) %>%
coo_slide(ldk=1) %>% efourier(norm=FALSE) %>% PCA() %>%
# now the LDA and its summary
LDA(~aut) %>% classification_metrics()

#> iteration: 1	gain:44481

#> iteration: 2	gain:37.931

#> iteration: 3	gain:11.298

#> iteration: 4	gain:0.14521

#> 'nb.h' set to 7 (99% harmonic power)

#> 10 PC retained

#> $accuracy
#> [1] 0.775
#> 
#> $macro_prf
#> # A tibble: 8 x 3
#>   precision recall    f1
#>       <dbl>  <dbl> <dbl>
#> 1     0.867  0.867 0.867
#> 2     0.75   0.8   0.774
#> 3     0.583  0.467 0.519
#> 4     0.926  0.833 0.877
#> 5     0.862  0.833 0.847
#> 6     0.839  0.867 0.852
#> 7     0.871  0.9   0.885
#> 8     0.528  0.633 0.576
#> 
#> $macro_avg
#> # A tibble: 1 x 3
#>   avg_precision avg_recall avg_f1
#>           <dbl>      <dbl>  <dbl>
#> 1         0.778      0.775  0.775
#> 
#> $ova
#> $ova$ced
#>         classified
#> actual   ced others
#>   ced     26      4
#>   others   4    206
#> 
#> $ova$jeya
#>         classified
#> actual   jeya others
#>   jeya     24      6
#>   others    8    202
#> 
#> $ova$mat
#>         classified
#> actual   mat others
#>   mat     14     16
#>   others  10    200
#> 
#> $ova$ponnu
#>         classified
#> actual   ponnu others
#>   ponnu     25      5
#>   others     2    208
#> 
#> $ova$remi
#>         classified
#> actual   remi others
#>   remi     25      5
#>   others    4    206
#> 
#> $ova$rom
#>         classified
#> actual   rom others
#>   rom     26      4
#>   others   5    205
#> 
#> $ova$ruks
#>         classified
#> actual   ruks others
#>   ruks     27      3
#>   others    4    206
#> 
#> $ova$vince
#>         classified
#> actual   vince others
#>   vince     19     11
#>   others    17    193
#> 
#> 
#> $ova_sum
#>           classified
#> actual     relevant others
#>   relevant      186     54
#>   others         54   1626
#> 
#> $kappa
#> [1] 0.7428571
#>