Combines Momocs objects

Combine Coo objects after a slicing, either manual or using slice or chop. Note that on Coo object, it combines row-wise (ie, merges shapes as a c would do) ; but on Coe it combines column-wise (merges coefficients). In the latter case, Coe must have the same number of shapes (not necessarily the same number of coefficients). Also the $fac of the first Coe is retrieved. A separate version may come at some point.

combine(...)

Arguments

...
a list of Out(Coe), Opn(Coe), Ldk objects (but of the same class)

Note

Note that the order of shapes or their coefficients is not checked, so anything with the same number of rows will be merged.

See also

Other handling functions: arrange, at_least, chop, dissolve, filter, mutate, rename, rescale, rm_harm, rm_uncomplete, rw_fac, sample_frac, sample_n, select, slice, subset.Coo, table

Examples

w <- filter(bot, type=="whisky")
b <- filter(bot, type=="beer")
combine(w, b)

#> An Out object with: 
#>  - $coo: 40 outlines (162 +/- 21 coordinates, all unclosed)
#>  - $fac: 1 classifier:
#>      'type' (factor 2): whisky, beer.

# or, if you have many levels
bot_s <- chop(bot, type)
bot_s$whisky

#> An Out object with: 
#>  - $coo: 20 outlines (158 +/- 23 coordinates, all unclosed)
#>  - $fac: 1 classifier:
#>      'type' (factor 2): beer, whisky.

# note that you can apply something (single function or a more
# complex pipe) then combine everyone, since combine also works on lists
# eg:
# bot_s2 <- lapply(bot_s, efourier, 10)
# bot_sf <- combine(bot_s2)

# pipe style
lapply(bot_s, efourier, 10) %>% combine()

#> An OutCoe object [ combined: efourier + efourier analyses ]
#> --------------------
#>  - $coe: 20 outlines described, and (total) 80 coefficients
#>  - $fac: 1 classifier:
#>      'type' (factor 2): beer, whisky.