Calculates the Haralick's circularity of a shape

coo_circularity calculates the 'circularity measure'. Also called 'compactness' and 'shape factor' sometimes. coo_circularityharalick calculates Haralick's circularity which is less sensible to digitalization noise than coo_circularity. coo_circularitynorm calculates 'circularity', also called compactness and shape factor, but normalized to the unit circle.

coo_circularity(coo)

# S3 method for default
coo_circularity(coo)

# S3 method for Coo
coo_circularity(coo)

coo_circularityharalick(coo)

# S3 method for default
coo_circularityharalick(coo)

# S3 method for Coo
coo_circularityharalick(coo)

coo_circularitynorm(coo)

# S3 method for default
coo_circularitynorm(coo)

# S3 method for Coo
coo_circularitynorm(coo)

Arguments

coo
a matrix of (x; y) coordinates or any Coo

Source

Rosin PL. 2005. Computing global shape measures. Handbook of Pattern Recognition and Computer Vision. 177-196.

Value

numeric for single shapes, list for Coo of the corresponding circularity measurement.

See also

Other coo_ descriptors: coo_angle_edge1, coo_angle_edges, coo_angle_tangent, coo_area, coo_boundingbox, coo_chull, coo_convexity, coo_eccentricity, coo_elongation, coo_length, coo_lw, coo_rectangularity, coo_rectilinearity, coo_solidity, coo_width

Examples

# coo_circularity
bot[1] %>% coo_circularity()

#> [1] 26.26463

bot %>%
    slice(1:5) %>% # for speed sake only
    coo_circularity

#> $brahma
#> [1] 26.26463
#> 
#> $caney
#> [1] 25.60553
#> 
#> $chimay
#> [1] 20.83278
#> 
#> $corona
#> [1] 27.61134
#> 
#> $deusventrue
#> [1] 25.75573
#> 

# coo_circularityharalick
bot[1] %>% coo_circularityharalick()

#> [1] 2.320493

bot %>%
    slice(1:5) %>% # for speed sake only
    coo_circularityharalick

#> $brahma
#> [1] 2.320493
#> 
#> $caney
#> [1] 2.374045
#> 
#> $chimay
#> [1] 2.935174
#> 
#> $corona
#> [1] 2.261573
#> 
#> $deusventrue
#> [1] 2.397828
#> 

# coo_circularitynorm
bot[1] %>% coo_circularitynorm()

#> [1] 2.090073

bot %>%
    slice(1:5) %>% # for speed sake only
    coo_circularitynorm

#> $brahma
#> [1] 2.090073
#> 
#> $caney
#> [1] 2.037623
#> 
#> $chimay
#> [1] 1.65782
#> 
#> $corona
#> [1] 2.197241
#> 
#> $deusventrue
#> [1] 2.049576
#>