Calculates perimeter and variations

coo_perim calculates the perimeter; coo_perimpts calculates the euclidean distance between every points of a shape; coo_perimcum does the same and calculates and cumulative sum.

coo_perimpts(coo)

# S3 method for default
coo_perimpts(coo)

# S3 method for Coo
coo_perimpts(coo)

coo_perimcum(coo)

# S3 method for default
coo_perimcum(coo)

# S3 method for Coo
coo_perimcum(coo)

coo_perim(coo)

# S3 method for default
coo_perim(coo)

# S3 method for Coo
coo_perim(coo)

Arguments

coo
matrix of (x; y) coordinates or any Coo

Value

numeric the distance between every point or a list of those.

See also

Other coo_ utilities: coo_aligncalliper, coo_alignminradius, coo_alignxax, coo_align, coo_baseline, coo_bookstein, coo_boundingbox, coo_calliper, coo_centdist, coo_center, coo_centpos, coo_close, coo_down, coo_dxy, coo_enlarge, coo_extract, coo_flipx, coo_force2close, coo_interpolate, coo_jitter, coo_left, coo_nb, coo_rev, coo_right, coo_rotatecenter, coo_rotate, coo_sample_prop, coo_samplerr, coo_sample, coo_scale, coo_shearx, coo_slice, coo_slidedirection, coo_slidegap, coo_slide, coo_smoothcurve, coo_smooth, coo_template, coo_trans, coo_trimbottom, coo_trimtop, coo_trim, coo_up, is_clockwise, is_closed, is_equallyspacedradii

Examples

# for speed sake
b1 <- coo_sample(bot[1], 12)
b5 <- bot %>% slice(1:5) %>% coo_sample(12)

# coo_perim
coo_perim(b1)

#> [1] 2140.62

coo_perim(b5)

#> $brahma
#> [1] 2140.62
#> 
#> $caney
#> [1] 1942.307
#> 
#> $chimay
#> [1] 1354.083
#> 
#> $corona
#> [1] 1555.883
#> 
#> $deusventrue
#> [1] 1768.427
#> 

# coo_perimpts
coo_perimpts(b1)

#>  [1] 201.6829 212.0424 168.6001 198.8291 189.3198 221.2261 190.0026 206.7293
#>  [9] 192.2134 137.2079 222.7667

b5 %>% coo_perimpts()

#> $brahma
#>  [1] 201.6829 212.0424 168.6001 198.8291 189.3198 221.2261 190.0026 206.7293
#>  [9] 192.2134 137.2079 222.7667
#> 
#> $caney
#>  [1] 192.0000 181.0000 161.9784 175.0457 192.1276 171.0000 186.1317 191.7707
#>  [9] 175.3454 134.0149 181.8928
#> 
#> $chimay
#>  [1] 131.03435 123.06502 113.22544 128.99612 119.94165 131.03435 140.24621
#>  [8] 129.44883 118.37652  97.65244 121.06197
#> 
#> $corona
#>  [1] 155.0000 143.0315 131.2288 127.6323 145.0034 155.0000 154.9322 140.3567
#>  [9] 151.3275 112.9292 139.4417
#> 
#> $deusventrue
#>  [1] 175.1713 163.3065 147.1224 151.2415 175.5591 162.3730 182.3623 171.8430
#>  [9] 157.6230 131.8218 150.0033
#> 

# coo_perimcum
b1 %>% coo_perimcum()

#>  [1]    0.0000  201.6829  413.7254  582.3255  781.1546  970.4744 1191.7005
#>  [8] 1381.7032 1588.4324 1780.6459 1917.8537 2140.6204

b5 %>% coo_perimcum()

#> $brahma
#>  [1]    0.0000  201.6829  413.7254  582.3255  781.1546  970.4744 1191.7005
#>  [8] 1381.7032 1588.4324 1780.6459 1917.8537 2140.6204
#> 
#> $caney
#>  [1]    0.0000  192.0000  373.0000  534.9784  710.0241  902.1517 1073.1517
#>  [8] 1259.2833 1451.0540 1626.3994 1760.4143 1942.3072
#> 
#> $chimay
#>  [1]    0.0000  131.0343  254.0994  367.3248  496.3209  616.2626  747.2969
#>  [8]  887.5431 1016.9920 1135.3685 1233.0209 1354.0829
#> 
#> $corona
#>  [1]    0.0000  155.0000  298.0315  429.2603  556.8926  701.8960  856.8960
#>  [8] 1011.8282 1152.1849 1303.5124 1416.4416 1555.8833
#> 
#> $deusventrue
#>  [1]    0.0000  175.1713  338.4778  485.6002  636.8417  812.4008  974.7739
#>  [8] 1157.1361 1328.9791 1486.6021 1618.4239 1768.4272
#>