Calculates the rectilinearity of a shape

As proposed by Zunic and Rosin (see below). May need some testing/review.

Usage

coo_rectilinearity(coo)

Source

Zunic J, Rosin PL. 2003. Rectilinearity measurements for polygons. IEEE Transactions on Pattern Analysis and Machine Intelligence 25: 1193-1200.

Arguments

coo

a matrix of (x; y) coordinates or any Coo

Value

numeric for a single shape, list for Coo

Note

due to the laborious nature of the algorithm (in nb.pts^2), and of its implementation, it may be very long to compute.

See also

Examples

bot[1] %>%
    coo_sample(32) %>% # for speed sake only
    coo_rectilinearity
#> [1] 0.3539899

bot %>%
    slice(1:3) %>% coo_sample(32) %>% # for speed sake only
    coo_rectilinearity
#> $brahma
#> [1] 0.3539899
#> 
#> $caney
#> [1] 0.3751378
#> 
#> $chimay
#> [1] 0.3597856
#>