Brew and serve fac from Momocs object

Ease various specifications for fac specification when passed to Momocs objects. Intensively used (internally).

fac_dispatcher(x, fac)

Arguments

x
a Momocs object (any Coo, Coe, PCA, etc.)

fac
a specification to extract from fac

Value

a prepared factor (or a numeric). See examples

Details

fac can be:

a factor, passed on the fly

a column id from $fac

a column name from fac

a formula (preferred) in the form: ~column_name (from $fac, no quotes)

See also

Other handling functions: arrange, at_least, chop, combine, dissolve, filter, mutate, rescale, rm_harm, rm_uncomplete, rw_fac, sample_frac, sample_n, select, slice, subsetize

Examples

bot <- mutate(bot, s=rnorm(40), fake=factor(rep(letters[1:4], 10)))

# factor, on the fly
fac_dispatcher(bot, factor(rep(letters[1:4], 10)))

#>  [1] a b c d a b c d a b c d a b c d a b c d a b c d a b c d a b c d a b c d a b
#> [39] c d
#> Levels: a b c d

# column id
fac_dispatcher(bot, 1)

#>  type1  type2  type3  type4  type5  type6  type7  type8  type9 type10 type11 
#> whisky whisky whisky whisky whisky whisky whisky whisky whisky whisky whisky 
#> type12 type13 type14 type15 type16 type17 type18 type19 type20 type21 type22 
#> whisky whisky whisky whisky whisky whisky whisky whisky whisky   beer   beer 
#> type23 type24 type25 type26 type27 type28 type29 type30 type31 type32 type33 
#>   beer   beer   beer   beer   beer   beer   beer   beer   beer   beer   beer 
#> type34 type35 type36 type37 type38 type39 type40 
#>   beer   beer   beer   beer   beer   beer   beer 
#> Levels: beer whisky

# column name
fac_dispatcher(bot, "type")

#>  type1  type2  type3  type4  type5  type6  type7  type8  type9 type10 type11 
#> whisky whisky whisky whisky whisky whisky whisky whisky whisky whisky whisky 
#> type12 type13 type14 type15 type16 type17 type18 type19 type20 type21 type22 
#> whisky whisky whisky whisky whisky whisky whisky whisky whisky   beer   beer 
#> type23 type24 type25 type26 type27 type28 type29 type30 type31 type32 type33 
#>   beer   beer   beer   beer   beer   beer   beer   beer   beer   beer   beer 
#> type34 type35 type36 type37 type38 type39 type40 
#>   beer   beer   beer   beer   beer   beer   beer 
#> Levels: beer whisky

# same, numeric case
fac_dispatcher(bot, "s")

#>         s1         s2         s3         s4         s5         s6         s7 
#> -0.1958989  1.1775844  0.6834736 -1.2754967  0.6379564 -1.3775896 -0.5983108 
#>         s8         s9        s10        s11        s12        s13        s14 
#>  1.2109204 -2.2510452 -1.7790142  1.3013727 -0.8147928  1.2437070 -0.1682502 
#>        s15        s16        s17        s18        s19        s20        s21 
#>  0.4277757  0.8132789 -0.6512119 -0.3045909 -0.4150972  2.8160843  0.1261471 
#>        s22        s23        s24        s25        s26        s27        s28 
#>  0.4728004 -0.3407535 -0.2417906  1.3787547 -0.3388837  0.0201363  0.3769622 
#>        s29        s30        s31        s32        s33        s34        s35 
#> -0.4317237  1.9590642 -1.4284596  2.0112930 -0.3515919  1.3571196 -1.9991774 
#>        s36        s37        s38        s39        s40 
#>  0.9560806  0.8764313 -1.2712170 -0.7683239  0.1935248 

# formula interface
fac_dispatcher(bot, ~type)

#>  [1] whisky whisky whisky whisky whisky whisky whisky whisky whisky whisky
#> [11] whisky whisky whisky whisky whisky whisky whisky whisky whisky whisky
#> [21] beer   beer   beer   beer   beer   beer   beer   beer   beer   beer  
#> [31] beer   beer   beer   beer   beer   beer   beer   beer   beer   beer  
#> Levels: beer whisky

# formula interface + interaction on the fly
fac_dispatcher(bot, ~type+fake)

#>  [1] whisky_a whisky_b whisky_c whisky_d whisky_a whisky_b whisky_c whisky_d
#>  [9] whisky_a whisky_b whisky_c whisky_d whisky_a whisky_b whisky_c whisky_d
#> [17] whisky_a whisky_b whisky_c whisky_d beer_a   beer_b   beer_c   beer_d  
#> [25] beer_a   beer_b   beer_c   beer_d   beer_a   beer_b   beer_c   beer_d  
#> [33] beer_a   beer_b   beer_c   beer_d   beer_a   beer_b   beer_c   beer_d  
#> Levels: beer_a beer_b beer_c beer_d whisky_a whisky_b whisky_c whisky_d