Sample n shapes

Sample n shapes from a Momocs object. See examples and ?dplyr::sample_n.

sample_n(tbl, size, replace, fac, ...)

Arguments

tbl
a Momocs object (Coo, Coe)

size
numeric how many shapes should we sample

replace
logical whether sample should be done with ot without replacement

fac
a column name if a $fac is defined; size is then applied within levels of this factor

...
additional arguments to dplyr::sample_n and to maintain generic compatibility

See also

Examples

# samples 5 bottles no matter their type
sample_n(bot, 5)
#> Out (outlines)
#>   - 5 outlines, 163 +/- 26 coords (in $coo)
#>   - 2 classifiers (in $fac): 
#> # A tibble: 5 × 2
#>   type   fake 
#>   <fct>  <fct>
#> 1 whisky a    
#> 2 whisky b    
#> 3 beer   d    
#> 4 whisky b    
#> 5 whisky b    
#>   - also: $ldk
# 5 bottles of beer and of whisky
table(sample_n(bot, 5, fac="type")$type)
#> 
#>   beer whisky 
#>      5      5 
# many repetitions
table(names(sample_n(bot, 400, replace=TRUE)))
#> 
#>          amrut    ballantines         brahma      bushmills          caney 
#>             11             12              4             15             12 
#>         chimay         chivas         corona        dalmore    deusventrue 
#>             13             11             12             11              7 
#>          duvel   famousgrouse   franziskaner    glendronach   glenmorangie 
#>              5             13              9              5             13 
#>     grimbergen        guiness   highlandpark     hoegardeen    jackdaniels 
#>              8              9             13              4              8 
#>             jb  johnniewalker        jupiler     kingfisher       latrappe 
#>             11              8             11              9              7 
#> lindemanskriek       magallan     makersmark    nicechouffe           oban 
#>             10             10              6             14             12 
#>     oldpotrero     pecheresse      redbreast   sierranevada         tamdhu 
#>              9             12              8             10             19 
#>     tanglefoot          tauro      westmalle     wildturkey         yoichi 
#>              8             11             12              9              9